complexvalue valueholders now return NULL when
they are json encoded for example. Before it was string, which doesn't
make sense in the case of objects. Returning empty arrays doesn't make
sense for the native serialization as well as that would be rejected on
recreation of the entity. With NULL the attribute's value falls back to
the NullValue of the attribute which is NULL for the existing complexvalue
attributes.