add HasComplexValueInterface so applications know
whether an attribute has a ComplexValue extending value in their
valueholder or not. This should make it easier to cleanup submitted empty
form data for those attributes. See GeoPoint class as an example where one
may not want to submit an input form with 'lon => '' and 'lat' => '' and
setting the lon/lat to 0 is not the correct solution either – especially
when there's no default value set on the attribute.