fix: two timestamp values always treated as equal
Fixes the TimestampValueHolder::sameValueAs() comparison of two (object) values.

The Timestamp/DateAttribute valueholders accept NULL as a valid value to
basically unset the set date. The code for that has been simplified
(see apply() method) and now expects the execute() method to always set a
sanitized value as a side effect instead of afterwards guessing and
comparing NULL or empty string (which may be used to unset the value as
well).

refs #12