introduce filesize, filename and mimetype on image
value object and attribute (asset vo as well) and add a new
SanitizedFilenameRule that is used to validate the filename property. The
filename value will be stripped of various problematic characters by
default to minimize the risk when uploaded file's client provided
filenames are stored in this property. The sanitization doesn't mean that
it's a good idea to store the user provided filenames and it naturally
doesn't mean that the name should be used for downloading the asset/image.