GeoPoint attribute now accepts NULL & '' values
This allows resetting the valid value of GeoPoint attributes to NULL.
Allowing empty strings to be treated as NULL is done on purpose to
make it easier to reset lon/lat values to default/null values.

Prior to this change the rule did not accept NULL or empty string as valid
input and the value holder implementation did not check for equal values
that might be NULL or empty string.