add handling of Null Island (0,0) as GeoPoint
The attribute can now specify via rule option 'null_island_as_null'
whether the (0,0) coordinate is accepted as a valid value or whether it
leads to the attribute value holder value being set to NULL.

By default Null Island is treated as a geocoding failure and thus the
value of the attribute is set to the null value. By setting the above
option the explicit setting of (0,0) as coordinates is accepted.