unscientific benchmarking of TextRule/createEntity
The toBoolean() method with it's filter_var method seems to be a low
hanging fruit while not losing much functionality.

For ArticleType->createEntity not using toBoolean() for getOption() is
about 7% faster (1000 articles created).

For TextRule apply the speedup of not using toBoolean() for getOption() is
about 18% faster on short text and ~5% faster on a 20x longer text.
When additionally to not using toBoolean using the trim() method instead of
the UTF8 regex we use atm the speedup is about 35% on short text and about
57% on a 20x longer text (still not that long).

Test runs were being done in a VM with PHP 5.6. Host system is a i7-2620M
CPU @ 2.70GHz.

refs #14