remove default zeroing of time for DateAttribute
this leads to wrong interpretation of dates (off-by-one) as the time is
changed. The native format used still contains the time information to
provide the timezone information for consumers. A short format like
'yyyy-MM-dd+00:00' is not being used even though php likes it other
parties interpreting that value might not like it. One example are api
consumers that get that native value or even elasticsearch which wants to
see something like "yyyy-MM-ddT+00:00". That would be parsed as military
Tango time though by php. That's why we sent the time info. Meh.
Suggestions are welcome. Applications can reformat according to their
needs on their system boundaries.

Future improvement suggestion: internally store the timezone of the given
values to have them available later on. This helps in migrating stuff when
eg.g the timezone db changes.

refs #28