possible BC: fix EntityMap::hasItem() method and
related functionality. Add tests for it. Prior to the the identifier of
the entities wasn't used for comparison whether an item is already present
in the EntityMap. As the identity in our case is specified by the
identifier value this leads to hasKey($entity->getIdentifier()) and
hasItem($entity) behaving differently as the comparison for the hasItem()
method fails for other instances of an entity with the same identifier.